individual directories for CFEnv processing components
