+++ amqp-cfenv-processor/src/test/resources/messages-rabbit-vcap-services.json
