+++ amqp-cfenv-processor/src/test/resources/mockito-extensions/org.mockito.plugins.MockMaker
