--- data-cfenv-processor/pom.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>com.ibm.cloud</groupId>
-        <artifactId>cfenv-processor-parent</artifactId>
-        <version>0.0.1.BUILD-SNAPSHOT</version>
-    </parent>
-
-    <groupId>com.ibm.cloud</groupId>
-    <artifactId>data-cfenv-processor</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
-    <name>data-cfenv-processor</name>
-    <description>A CfEnvProcessor Implementation for Cloudant</description>
-    <properties>
-        <version.spring-rabbit>2.2.1.RELEASE</version.spring-rabbit>
-    </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>com.ibm.cloud</groupId>
-            <artifactId>ssl-context-spring-boot-starter</artifactId>
-            <version>0.0.1-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>com.ibm.cloud</groupId>
-            <artifactId>bean-customizer-spring-boot</artifactId>
-            <version>0.0.1-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.amqp</groupId>
-            <artifactId>spring-rabbit</artifactId>
-            <version>${version.spring-rabbit}</version>
-        </dependency>
-
-    </dependencies>
-
-</project>
