--- data-cfenv-processor/src/main/java/com/ibm/cfenv/spring/boot/data/amqp/AmqpSSLContextConfiguration.java
@@ -1,16 +1,0 @@
-package com.ibm.cfenv.spring.boot.data.amqp;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import javax.net.ssl.SSLContext;
-import java.util.Map;
-
-@Configuration
-public class AmqpSSLContextConfiguration {
-
-    @Bean
-    public AmqpSSLContextBeanCustomizer beanCustomizer(Map<String, SSLContext> sslContexts) {
-        return new AmqpSSLContextBeanCustomizer(sslContexts);
-    }
-}
