--- data-cfenv-processor/src/test/java/com/ibm/cfenv/spring/boot/data/CloudantCfEnvProcessorApplicationTests.java
@@ -1,14 +1,0 @@
-package com.ibm.cfenv.spring.boot.data;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-public class CloudantCfEnvProcessorApplicationTests {
-
-    @Test
-    public void contextLoads() {
-    }
-
-}
