+++ db2-cfenv-processor/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,1 @@
+io.pivotal.cfenv.spring.boot.CfEnvProcessor=com.ibm.cfenv.spring.boot.db2.DB2CfEnvProcessor
