+++ pom.xml
@@ -44,7 +44,9 @@
     </dependencies>
 
     <modules>
-        <module>data-cfenv-processor</module>
+        <module>amqp-cfenv-processor</module>
+        <module>cloudant-cfenv-processor</module>
+        <module>db2-cfenv-processor</module>
         <module>mongodb-cfenv-processor</module>
         <module>watson-cfenv-processor</module>
         <module>bean-customizer-spring-boot</module>
